# ipx/api/handler/v2/transaction_types.go
package v2

import (
	"encoding/hex"
	"errors"
	"math/big"
	"strings"

	"github.com/andantan/evmlab/core/types"
	"github.com/ethereum/go-ethereum/common"
)

type BuildNativeLegacyTransactionRequest struct {
	From  string `json:"from"`
	To    string `json:"to"`
	Value string `json:"value"`

	value *big.Int
	from  *types.Address
	to    *types.Address
}

func (r *BuildNativeLegacyTransactionRequest) ValidateRequest() error {
	r.From = strings.TrimSpace(r.From)
	if r.From == "" {
		return errors.New("from is required")
	}
	if !common.IsHexAddress(r.From) {
		return errors.New("from: invalid address")
	}
	r.from = types.NewAddress(common.HexToAddress(r.From))

	r.To = strings.TrimSpace(r.To)
	if r.To == "" {
		return errors.New("to is required")
	}
	if !common.IsHexAddress(r.To) {
		return errors.New("to: invalid address")
	}
	r.to = types.NewAddress(common.HexToAddress(r.To))

	r.Value = strings.TrimSpace(r.Value)
	if r.Value == "" {
		return errors.New("value is required")
	}
	v, ok := new(big.Int).SetString(r.Value, 10)
	if !ok {
		return errors.New("value: must be a decimal wei amount")
	}
	r.value = v

	return nil
}

func (r *BuildNativeLegacyTransactionRequest) FromAddr() *types.Address { return r.from }
func (r *BuildNativeLegacyTransactionRequest) ToAddr() *types.Address   { return r.to }
func (r *BuildNativeLegacyTransactionRequest) Amount() *big.Int         { return r.value }

type BuildNativeLegacyTransactionResponse struct {
	UnsignedRLP string `json:"unsigned_rlp"`
}

func NewBuildNativeLegacyTransaction(data []byte) *BuildNativeLegacyTransactionResponse {
	return &BuildNativeLegacyTransactionResponse{
		UnsignedRLP: "0x" + hex.EncodeToString(data),
	}
}

type BuildNativeEIP1559TransactionRequest struct {
	From  string `json:"from"`
	To    string `json:"to"`
	Value string `json:"value"`

	v *big.Int
	f *types.Address
	t *types.Address
}

func (r *BuildNativeEIP1559TransactionRequest) ValidateRequest() error {
	r.From = strings.TrimSpace(r.From)
	if r.From == "" {
		return errors.New("from is required")
	}
	if !common.IsHexAddress(r.From) {
		return errors.New("from: invalid address")
	}
	r.f = types.NewAddress(common.HexToAddress(r.From))

	r.To = strings.TrimSpace(r.To)
	if r.To == "" {
		return errors.New("to is required")
	}
	if !common.IsHexAddress(r.To) {
		return errors.New("to: invalid address")
	}
	r.t = types.NewAddress(common.HexToAddress(r.To))

	r.Value = strings.TrimSpace(r.Value)
	if r.Value == "" {
		return errors.New("value is required")
	}
	v, ok := new(big.Int).SetString(r.Value, 10)
	if !ok {
		return errors.New("value: must be a decimal wei amount")
	}
	r.v = v

	return nil
}

func (r *BuildNativeEIP1559TransactionRequest) FromAddr() *types.Address { return r.f }
func (r *BuildNativeEIP1559TransactionRequest) ToAddr() *types.Address   { return r.t }
func (r *BuildNativeEIP1559TransactionRequest) Amount() *big.Int         { return r.v }

type BuildNativeEIP1559TransactionResponse struct {
	UnsignedRLP string `json:"unsigned_rlp"`
}

func NewBuildNativeEIP1559Transaction(data []byte) *BuildNativeEIP1559TransactionResponse {
	return &BuildNativeEIP1559TransactionResponse{
		UnsignedRLP: "0x" + hex.EncodeToString(data),
	}
}