# ipx/docs/swagger.yaml
basePath: /
definitions:
...
  v2.BuildNativeEIP1559TransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v2.BuildNativeEIP1559TransactionResponse:
    properties:
      unsigned_rlp:
        type: string
    type: object
  v2.BuildNativeLegacyTransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v2.BuildNativeLegacyTransactionResponse:
    properties:
      unsigned_rlp:
        type: string
    type: object
  v3.BuildNativeEIP1559TransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        description: wei as decimal string
        type: string
    type: object
  v3.BuildNativeEIP1559TransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v3.BuildNativeLegacyTransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        description: wei as decimal string
        type: string
    type: object
  v3.BuildNativeLegacyTransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
...
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/v2/transfers/native/eip1559:
    post:
      consumes:
      - application/json
      description: Fetches chain state (chainID, nonce, fees) and returns the unsigned
        RLP-encoded EIP-1559 transfer transaction
      parameters:
      - description: BuildNativeEIP1559Transaction request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v2.BuildNativeEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.BuildNativeEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build unsigned EIP-1559 transfer tx
      tags:
      - transfer
  /evm/v2/transfers/native/legacy:
    post:
      consumes:
      - application/json
      description: Fetches chain state (chainID, nonce, gas price) and returns the
        unsigned RLP-encoded EIP-155 legacy transfer transaction
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v2.BuildNativeLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.BuildNativeLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build unsigned legacy transfer tx
      tags:
      - transfer
  /evm/v3/transactions/native/eip1559:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-1559 native transfer,
        returns unsigned_rlp, signed_rlp, and tx_hash
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v3.BuildNativeEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.BuildNativeEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build and sign EIP-1559 transfer tx
      tags:
      - transaction
  /evm/v3/transactions/native/legacy:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-155 legacy native
        transfer, returns unsigned_rlp, signed_rlp, and tx_hash
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v3.BuildNativeLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.BuildNativeLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build and sign legacy transfer tx
      tags:
      - transaction
...
swagger: "2.0"